Add debug="when_hidden" option.
By setting the shortcode variable `debug` to `"when_hidden"`, the
debugging statements will only be added to the page when the content is
hidden by the plugin.

By doing this, anyone that can edit the page can have a clear view of
what will be published and when, without having to deal with multiple
debugging statements for already displayed content.